chore(collector-rule): move from /schema to /api-clients/.../schema
