+++ apps/web/src/api-clients/inventory/collector-rule/schema/api-verbs/get.ts
