+++ apps/web/src/api-clients/inventory/collector-rule/schema/constant.ts
