+++ apps/web/src/services/asset-inventory/components/CollectorAdditionalRuleContent.vue
@@ -5,16 +5,17 @@ import type { TranslateResult } from 'vue-i18n';
 import { PLink, PI, PTooltip } from '@cloudforet/mirinae';
 import { isNotEmpty } from '@cloudforet/utils';
 
-import { useAllReferenceDataModel } from '@/query/resource-query/reference-model/use-all-reference-data-model';
 import {
     COLLECTOR_RULE_CONDITION_KEY,
     COLLECTOR_RULE_CONDITION_KEY_LABEL,
     COLLECTOR_RULE_CONDITION_OPERATOR_LABEL,
-} from '@/schema/inventory/collector-rule/constant';
-import type { CollectorRuleModel, AdditionalRuleAction } from '@/schema/inventory/collector-rule/model';
+} from '@/api-clients/inventory/collector-rule/schema/constant';
+import type { CollectorRuleModel, AdditionalRuleAction } from '@/api-clients/inventory/collector-rule/schema/model';
+import { useAllReferenceDataModel } from '@/query/resource-query/reference-model/use-all-reference-data-model';
 import { i18n as _i18n } from '@/translations';
 
 import { useReferenceRouter } from '@/router/composables/use-reference-router';
+
 
 interface Props {
     data?: CollectorRuleModel;
